perf: Prioritize non-weak streams over weak referenced streams. (#304)
If we prioritize pulling from the overflow streams (as it was before), it's possible for us to retain the weak references for a lot longer than is desirable as we will keep accessing them. It's therefore preferable to return a stream from the normal queue to allow the GC time to clean the weak references, and only return an overflow value if there is no other option.