Rename NetworkedTransform
New users (including me at one point) continue to be confused by NetworkedTransform and think it should be used in production. It is a great example for new users but is NOT good enough to use in production. New users should see that this is an example of how to network their transform, and not exactly how it should be used.

While NetworkedTransform does need some changes to be a better example, the least we can do for new users is rename NetworkedTransform to ExampleNetworkedTransform to help new users out.