ci: Update yamato CI to latest multi-package templates
Our custom CI configuration started failing after a change in the upm-ci tool. This change updates the CI configuration to use the standard templates for multi-package repositories. As a side effect, it also introduces a test/sample project where we can exercise the packages and put tests we don't want to ship with the packages themselves. Tests run against 2020.1 and trunk, on windows, mac, and ubuntu