ci: Fix syntax error on develop pull requests job
Templating code was commented out incorrectly and so was still being executed. This caused a UI bug on the Yamato website displaying duplicates of the same jobs.

This change also removes the direct package test dependencies from the "all tests" job, as they're pulled in through the "badges test trigger" dependency on that job.