Fix: Loopback RPCs in host mode (#460)
* fix: Host loopback

This fixes issue where Rpcs are not being invoked when in Host mode, but not for nested Rpcs.  (that is the next fix coming)

* fix: Host nested Rpcs

These are the first initial changes to fix the scenario where a Host has a ServerRpc being invoked which inside the Rpc it invokes a ClientRpc (etc.).

We now always have no less than 2 QueueHistoryFrames, when we are looping back Rpcs for a Host the looped back packet always gets placed on the next QueueHistory frame.  With 0 queue history, we will still have 2 frames to toggle between.

* refactor: removing loopback

Removed the loopback option since this now is the default mode when running in Host mode.
Cleaned up some legacy non-standard code as well.

* refactor:  syntax

Fixing minor syntax related issues.

* fix: Host should send to clients as well

This fixes the scenario where a Host would properly send to itself but would exclude the remaining clients.

* Fix: Host and Clients multiple streams

This fixes the scenario where you want to write to two streams but can only return one stream to the ILPP side of the RPC.
This will be fixed relatively soon with a mux'd bit serializer.

* fix: shutdown send exception

This prevents from trying to flush and send any remaining items if we are no longer connected.

* refactor: adjustments to bitserializer changes

Just making adjustments to the changes in parameters and names from the merged feature/serializable-types.
https://github.com/Unity-Technologies/com.unity.multiplayer.mlapi/tree/feature/serializable-types

* fix: minor adjustment for host loopback serverRpc

This fixes an offset issue with a host invoking a serverRpc.

* fix: renabling batching

Tested everything with batched RPCs disabled, now tested everything with batched RPCs enabled.
Host Loopback Unit test passed.
Server + Client  --> passed.
Host + Client --> passed.