fix: Using QueueHistoryFrame
Needed to refactor this approach by using the container, QueueHistoryFrame,  of  the next inbound frame in order to mirror what was written to the outbound.
Still building a test for this entire scenario, but currently this looks like a better approach until BitSerializers can multiplex writing when configured with more than one BitWriters.