fixes:
Potential Fix:  NetworkedVar write permissions was checking for IsServer but also checking for whether a client could write to the networkedvar in the same error catching scenario.  It looks like the intended logic was if you are "not the server" (i.e. a client) and you do not have write permissions then bail.

Updates to RPC Host mode loopback with additional fix for scenario where a queuehistory frame was considered invalid.