fix: fixes client not having all players upon first connecting
This update fixes the issue where the client would enter the lobby but not have all players listed until they clicked the ready button.  On the client side, I am pumping the server by notifying that I am "here" (even though the server knows this) but not ready.  The server, upon receiving the first status message, will apply the status and then send a notification out to all clients on the status of players in the lobby.