Update: Cameras, Rigid bodies, and more!
This is still WIP, but hiding players with some default items already in the samples is working, pausing appears to be working other than the SyncTransform continuing to update even when the rigid body's constraint is all.
Other than a better playerObject (tomorrow), this is a good point to get feedback.