fix and format:  Client camera and CR
Fixing some weird LFCR issues that started happening recently.
Fixed client camera not being synchronized when joining a game already in progress.
Fixed client not being able to exit.
Added a flag to the in game manager to tell it whether it should exit upon being the last player or not.