refactor: Players movement, collision, colors, and more
This update is close to the final version with potentially some additions to the graphics or the like in the next week.  The core framework for all of the systems are in place, players can join via lobby or join during game play without soft synch issues or issues with initializing positions.

Did some clean up of the comments, scenes (added exit button for lobby), etc.