fix: MTT-504
The connection approval process now guards against failure to set the NetworkConfig.ConnectionData as well as generates NetworkLog warning messages about the value not being set.
The connection approval process will continue to invoke the user's callback, which at that point the user can determine if a null value for the passed in connectionData is "ok" or not.

This fix includes an additional ConnectionApproveal sample that can be used to validate this fix.