refactor: Object Synchronization
This refactors how scene object synchronization is handled.  Primarily this puts the object synchronization found within the approval and scene transition process into one location which should help to assure that both cases are handled in the same way.

This also includes additional header information for network variable data written into the outbound stream as well as reading the added information on the client side from the inbound stream.