refactor and improve
Added additional test in scene transitioning that registers custom prefab handler via the NetworkObject as opposed to the GameObject.
Changed the mesh for each different scene transition.  Third scene uses pink spheres.
Expanded the MainMenuManager's scene registration to include a "scene type".
A scene can either be a Selectable or IncludeOnly (i.e. you want it included but don't want it to be an option in the menu).
Added additional code to handle these updates for runtime builds.
Updated the SampleScene to the new NetworkPrefab pattern.