Clean Up and New SamplesMenu
Did some cleaning up, adding exit buttons to all scenes, and added a SamplesMenu.
This also demonstrates how the MainMenuManager works.  If you open the build settings and then open the SamplesMenu you should only see the scenes included in the SamplesMenu's MainMenuManager.  If you then open the MTTManualTests scene you will notice only the scenes included in the MTTManualTests's MainMenuManager are enabled.  This provides a clean way to build for a specific set of scenes associated with a main menu.