feat!: Network Prefab Overrides, Inspector View, and the default Player Prefab (#749)
* NetworkPrefab registration and default PlayerPrefab

This is the first pass changes required to implement MTT-640.
The default player prefab is now a directly assigned value as opposed to a checkbox in the network prefab list.
The default player prefab has to also be included in the NetworkPrefab list (this is so the default player prefab can be overridden too)
NetworkPrefab entries can be overridden and directly assigned alternate prefabs for clients to spawn.
The source prefab (what the server uses) can be a prefab reference or a prefab hash.

* refactor: Naming, Migration, and PlayerPrefab

This update is the first pass of making this update as painless as possible to the user by "auto-handling" the v0.1.0 NetworkPrefab pattern as well as there is additional "user-bumper" code that assures the assigned default player prefab is always in the NetworkPrefab list.

* refactor: PlayerPrefab assignment

Removing the auto insert within the editor and doing this at runtime.
Added additional check to handle assigning the OverridingSourcePrefab if it is null and the Prefab is not.

* refactor: player prefab selection and blank new entries

This will prevent the scenario where a user might want to delete the default player prefab but also has it assigned within the NetworkPrefab list.

This assures that a newly created network prefab list entry is blank and that there are not two entries within the list that have the same global object id hash.

* style: comments

just adding a few comments

* style: property names and comments

Adjusted names of the new NetworkPrefab properties and added additional comments to better clarify some of the "trickery" happening that will help make this a "non-breaking" feature when it comes to existing v0.1.0 NetworkConfigs and the already selected default player prefab.

* refactor: Client Sync and Comments

This will prevent any future client side soft synchronization issues or related issues where only one (or a few) NetworkObjects are misconfigured and the client fails to synchronize (find or instantiate).  As opposed to completely breaking the rest of the pending NetworkObjects to be synchronized/instantiated, it will log an error and continue loading.

This also has some additional comments and updates to comments for better clarity.

* style: minor change

minor change in comments

* refactor: remove CreatePlayerPrefab

This removes the CreatePlayerPrefab property from the NetworkConfig and from the NetworkManagerEditor.  It is no longer needed as the NetworkConfig.PlayerPrefab property being something other than null tells us that the user wants to have MLAPI automatically spawn the players for them.

* refactor: cleaning up using

removed unused namespace references.

* style: PlayerPrefab tooltip

Added a tooltip for the new NetworkConfig.PlayerPrefab property

* style: spaces between slashes and comment text body

Added some spaces between the forward slashes and comment's text body

* fix: clearing network prefab links

This fixes an issue that can occur with serialization and NetworkPrefabOverrideLinks.
The policy is to ALWAYS clear this dictionary before building the links dictionary.

* refactor: NetworkConfig compatibility, NetworkPrefab, and Style

This removes the NetworkConfig compatibility checks that auto-upgraded a previous NetworkConfig setting to the new pattern.
Updated NetworkPrefab according to this change, isPlayer is no longer needed.
Added additional code to NetworkPrefab that will return a hash reflective of the override state.
Adjusted some comments.

* fix: wrong hash reference

Minor adjustment for the right GlobalObjectIdHash check.

* refactor and style

Made the NetworkPrefab internal as well as NetworkPrefabOverride.
Removed GetNetworkPrefabIndexOfHash and GetPrefabHashFromIndex as they were used in the old hash generation pattern.
Updated comments a bit.

* refactor: removing NetworkObject check

Removing two checks for the failure to spawn a NetworkObject locally.
This will be handled via PR-756:
https://github.com/Unity-Technologies/com.unity.multiplayer.mlapi/pull/756

* refactor: removing NetworkPrefab.Hash

This property is no longer being used and where it was used was replaced by the NetworkConfig.NetworkPrefabOverrideLinks

Co-authored-by: M. Fatih MAR <mfatihmar@gmail.com>