refactor, fix, and updated assets
A bunch of adjustments in the menu system.
Added command line handling.
This includes a fix for the "odd behavior" with GlobalObjectIdHash and the HandleApproval method in NetworkManager.
Unity Chan animator only requires one to press "c" once and it will animate until the end of the animation.
New SceneMenuReference for scenes that are sub-menus to the main menu.
Some bug fixes in various places (typically on exiting a scene).