Finalizing
Finished the scene and menu referencing pattern.
Now both menu systems use the same set of code and now there is a MenuReference and SceneReference that can be used in the MainMenuManager and SceneMenuManager (which really only define the reference type to use)
Minor tweaks in materials.
Remove some NetworkObject related debug logging.