refactor: scene object serialization consolidation and improved handling (#756)
* NetworkPrefab registration and default PlayerPrefab

This is the first pass changes required to implement MTT-640.
The default player prefab is now a directly assigned value as opposed to a checkbox in the network prefab list.
The default player prefab has to also be included in the NetworkPrefab list (this is so the default player prefab can be overridden too)
NetworkPrefab entries can be overridden and directly assigned alternate prefabs for clients to spawn.
The source prefab (what the server uses) can be a prefab reference or a prefab hash.

* refactor: Naming, Migration, and PlayerPrefab

This update is the first pass of making this update as painless as possible to the user by "auto-handling" the v0.1.0 NetworkPrefab pattern as well as there is additional "user-bumper" code that assures the assigned default player prefab is always in the NetworkPrefab list.

* refactor: PlayerPrefab assignment

Removing the auto insert within the editor and doing this at runtime.
Added additional check to handle assigning the OverridingSourcePrefab if it is null and the Prefab is not.

* refactor: player prefab selection and blank new entries

This will prevent the scenario where a user might want to delete the default player prefab but also has it assigned within the NetworkPrefab list.

This assures that a newly created network prefab list entry is blank and that there are not two entries within the list that have the same global object id hash.

* style: comments

just adding a few comments

* style: property names and comments

Adjusted names of the new NetworkPrefab properties and added additional comments to better clarify some of the "trickery" happening that will help make this a "non-breaking" feature when it comes to existing v0.1.0 NetworkConfigs and the already selected default player prefab.

* refactor: Client Sync and Comments

This will prevent any future client side soft synchronization issues or related issues where only one (or a few) NetworkObjects are misconfigured and the client fails to synchronize (find or instantiate).  As opposed to completely breaking the rest of the pending NetworkObjects to be synchronized/instantiated, it will log an error and continue loading.

This also has some additional comments and updates to comments for better clarity.

* style: minor change

minor change in comments

* refactor: remove CreatePlayerPrefab

This removes the CreatePlayerPrefab property from the NetworkConfig and from the NetworkManagerEditor.  It is no longer needed as the NetworkConfig.PlayerPrefab property being something other than null tells us that the user wants to have MLAPI automatically spawn the players for them.

* refactor: cleaning up using

removed unused namespace references.

* style: PlayerPrefab tooltip

Added a tooltip for the new NetworkConfig.PlayerPrefab property

* style: spaces between slashes and comment text body

Added some spaces between the forward slashes and comment's text body

* fix: clearing network prefab links

This fixes an issue that can occur with serialization and NetworkPrefabOverrideLinks.
The policy is to ALWAYS clear this dictionary before building the links dictionary.

* refactor: NetworkConfig compatibility, NetworkPrefab, and Style

This removes the NetworkConfig compatibility checks that auto-upgraded a previous NetworkConfig setting to the new pattern.
Updated NetworkPrefab according to this change, isPlayer is no longer needed.
Added additional code to NetworkPrefab that will return a hash reflective of the override state.
Adjusted some comments.

* refactor: Object Synchronization

This refactors how scene object synchronization is handled.  Primarily this puts the object synchronization found within the approval and scene transition process into one location which should help to assure that both cases are handled in the same way.

This also includes additional header information for network variable data written into the outbound stream as well as reading the added information on the client side from the inbound stream.

* refactor merge

Removing some code that got duplicated during a merge.

* minor update

* minor update

* minor update

* minor update

* style: Comments and Standards

Fixed method declaration standard issue.
Improved comments for network variable data size placeholder.
Using only NetworkWriter and NetworkReader as parameter values.

* Update com.unity.multiplayer.mlapi/Runtime/Core/NetworkObject.cs

Suggested update by mattwalsh-untity.
This just passes in the already obtained buffer as opposed to calling a method that returns the same buffer.
This is a cleaner implementation adjustment.

Co-authored-by: Matt Walsh <69258106+mattwalsh-unity@users.noreply.github.com>

* refactor: Naming updates, comments, an d logic

Refactored some of the names used to hopefully better clarify what is happening in this process.  Updated the logic and flow slightly to assure stream position alignment is accurate and for better clarity of what is happening.

* test: NetworkObject scene serialization test

This includes the test to validate that the primary purpose behind this PR, to skip past an invalid NetworkObject during scene object synchronization, is functioning as expected.

* refactor test

This is an improved NetworkObjectSceneSerializationFailure unit test.
It adds random NetworkVariable data sizes to assure all stream offsets are being calculated properly.  It randomly adds invalid NetworkObjects.

* style

Minor style related changes. Adjusted comment and removed LF.

* style: invalidNetworkObjectFrequency

comment update for invalidNetworkObjectFrequency

* style: spelling

Fixed some spelling issues.

Co-authored-by: M. Fatih MAR <mfatihmar@gmail.com>
Co-authored-by: Matt Walsh <69258106+mattwalsh-unity@users.noreply.github.com>