test: connection approval
This is the manual test for connection approval.
It allows one to simulate connection approval failure as well as override the default player via in-game UI checkbox elements for ease of testing.
This includes the SceneReference for the new manual test scene and this SceneReference is included in the ManualTests menu.

Acceptance Tests Covered:
Connection approval and failure works as expected.
Players can have their default player replaced via the connection approval process.