refactor:
This is an updated improvement upon the initial NetworkVariableTestComponent and associated files.
Added the NetworkVariableHelper class that may be useful in the future.
Updated RpcQueueTests:
Reduced the wait time to get a slightly faster test time.
Used ordering to assure the RpcQueueContainerClass test is always ran last (of the 3 tests).
Updated some of the comments.