test: scene transitioning within testproject (#815)
* fix: Set singleton if needed and style update

Sometimes unit tests can run into a scenario where the singleton has yet to be assigned by the time the next unit test has started and instantiated a new NetworkManager instance.  This fixes that issue.

This also has some minor style updates to the comments.

* refactor: remove game objects after test finishes

This cleans up the network objects instantiated at the end of the test so they don't hang around in the test runner temporary scene.

This includes minor comment style updates.

* refactor: unit test scene transitioning

This is a temporary solution for switching scenes during unit tests using NetworkSceneManager.  When the NetworkSceneManager is overhauled, this will probably no longer be needed.

* test: scene transitioning

This validates the changes to the NetworkSceneManager as well as provides a working example of how to transition between scenes within testproject runtime unit tests.

* Revert "refactor: unit test scene transitioning"

This reverts commit 2a710459bd13b5b5bf94f231158b40df74c82ff3.

* refactor

Switching to a different pattern for unit test scene loading.

* refactor

Trying a different possible way to do this.

* style

updated comments.

* minor update

Co-authored-by: M. Fatih MAR <mfatihmar@gmail.com>