fix: Update package patch version to allow package registry re-publish (#874)
Updates the package version to 0.1.1 so we can publish it to Unity Package Registry.

Version 0.1.0 in the Unity Registry was published against an old commit and is missing several changes for that release. In order to publish a corrected version with all relevant changes, we have to bump the version number.