refactor and fix
Fixed some remaining checks (player prefab) as to whether the rigid body should be kinematic or not to prevent the mega spam.

During the revert changes of the NetworkTransform changes to better reflect what is in the develop branch, missed the else if.

Fixed a minor issue with the slider in the SceneTransitioning manual test to better visually reflect what the value is as well as handle the scenario if it starts at 0 boxes per second it needed to start the coroutine and if set to zero and the coroutine was running it terminates the coroutine.