SceneRegistration parenting (wip)
This update handles adding or removing the associated scenes marked to auto populate the build settings when a SceneRegistration asset is assigned or deleted to/from the NetworkManager.NetworkConfig.SceneRegistration property.

This also handles only adding or removing the SceneRegistrationEntries and AdditiveSceneEntries marked to auto populate the build settings with additional consideration of AdditiveSceneGroups nested within AdditiveSceneGroups.

Auto-populating the build settings portion is still a work in progress, but is getting closer to being completed.