fix
Went through to re-assign scene assets to some of the SceneRegistration assets that were older versions that had lost their entries (the scenes remained in the build settings but this needed to be updated anyway).
SceneRegistraitonEditor needed to make sure the entries list was allocated before trying to assign anything.
Adding or already having assigned the same SceneAsset that also references the SceneRegistration could cause the primary scene asset (i.e. one that has the NetworkManager instance that references it) do be deleted from the list.  This fixes that issue.