feat
Added new SCENE_EVENT that will handle "all things scene relative" with the only exception of the approval process.
That will need to be handled in a different PR as that requires a bunch of additional changes.
Focus for this branch is to get additive scene loading and unloading working.