feat
Improved the NetworkPrefabPool and NetworkPrefabPoolAdditive implementations to properly handle the various potential use case scenarios.  These changes assure that both the client and the server leverage from the pool.  This also takes into account NetworkManager defined NetworkPrefab overrides.

Added temporary LoadSceneEvent callback that notifies if an additive scene is being loaded or unloaded.

Updated the NetworkPrefabPoolAdditive sample to demonstrate how a user can assure their NetworkObjects are "spawned" in the same additive scene as their spawn generator.  This pattern will always destroy all NetworkObjects if the additive scene is unloaded.

Added sorting of NetworkObjects just prior to the server side serializing them in order to assure INetworkPrefabInstanceHandler implementations are registered first,  this is required in order to assure that the handler is registered before any NetworkObjects associated with it are spawned.