refactor and feat
Did some minor refactoring to make Load and Unload the only two methods that need to be called.  Removed the switch event related enum types since they are no longer used.

Added a re-synchronization step to handle edge case scenarios where a late joining client might miss destroy messages during the initial synchronization process.  This includes adding an additional DependentSceneName property to NetworkObject.  This lets the NetworkSceneManager know that a scene other than the NetworkObject's current scene is dependent upon this NetworkObject and to not instantiate it until that dependent scene is loaded.

Removed a bunch of user-side code (i.e. test project) from GenericNetworkObject that was required to handle the above edge case scenarios regarding scene dependencies.  Adjusted the NetworkPrefabPool and NetworkPrefabPoolAdditive to account for this addition, and removed a chunk of user-side code that would be required to handle this scenario.

Refactored SceneEventData to not implement INetworkSerializable which removes additional allocation for the NetworkSerializer during read and write operations.  Also adjusted the constructor to accept a NetworkManager instance to make it multi-instance compatible.

Set the SceneTransitioningBase1 NetworkManager config to not recycle NetworkObjectIds due to an issue in how they are recycled.  (topic for discussion)