feat
This creates a ScenesInBuild asset if one does not exist as well as populates it with the names of the scenes in the build to keep the same types of checks done within the NetworkSceneManager (i.e. does the scene index exist etc).
The ScenesInBuild asset can detect if it is moved around, however it is created in the root assets folder initially.  It also updates itself to the scenes in build list via the serialization process.
All NetworkManager instances reference this asset.  In turn, the NetworkSceneManager access the ScenesInBuild.Scenes list now.
Adjusted all unit tests and helpers for this update as well (except multiprocessor).
TestRunner passes all tests in editor and play modes.