FIx and Style
Fix for client side where during resync the GameObject was being destroyed but the spawn manager's lists tracking spawned NetworkObjects was not being updated.  The symptom was when a NetworkObjectId was recycled for a new NetoworkObject the client would still have that NetworkObjectId registered.

This also includes some style updates.