refactor and fix
Removing despawn replication and from places where they are no longer needed either due to the now proper message ordering.

Updated the GenericNetowrkObjectBehaviour such that clients disable the MeshRenderer component and then have a delayed re-enabling of it upon being spawned in order to avoid the latent update in the transform's position (removes the flickering of NetworkObjects on the client).

Added the ability to make the SwitchSceneHandler have a delayed "auto switch" in order to run cycled long burn runs to assure everything is working "as expected".

Fixed an access violation in StatsDisplay ( minor thing that was bugging me).