refactor!: NetworkPrefabHandler and INetworkPrefabInstanceHandler API & XmlDoc changes (#977)
* refactor and style

Renamed the two methods within INetworkPrefabInstanceHandler to better reflect what their function was and to avoid user confusion.

Improved comments and added additional XML Documentation for the NetworkPrefabHandler and associated INetworkPrefabInstanceHandler

* style

Minor updates on the ClientInstantiateOverride parameter names

* refactor and style

Renamed the interface methods.  Updated comments.
Refactored NetworkPrefabHandler to handle GitHub Issue 990 as well as made adjustments to be able to send the appropriate GlobalObjectIdHash as a host.
This also includes a public facing method in the NetworkManager to be able to lookup a NetworkManager defined NetworkPrefab Override for some associated cases relating to Host mode and NetworkPrefabHandlers and NetworkPrefab Overrides.
Made adjustments to the SceneTransitioning test and associated scenes and assets.

* refactor and style

Added the case where no prefab handler override is registered but there is still a host.
Added a lookup table to quickly get the proper source GlobalObjectIdHash value for the NetworkPrefab that has an override in the NetworkManager.
The first scene in the SceneTransitioningTest uses a NetworkManager defined prefab override, in host mode the host will pool while clients will not pool (no need since there is no way to override the instantiation) and the host will always send the source NetworkPrefab's GlobalObjectIdHash value as opposed to the target override's GlobalObjectIdHash.
Fixed Fatih's super-nit-pick :)

* fix

Removing an artifact property from a previous merge.

* fix

Fixing missing change to deregister and issue with despawning during fixed update.

* fix

Need to reset the position for the object when no handler exists otherwise the trigger will fire immediately and cause it to despawn immediately.

* refactor and fix

Removing despawn replication and from places where they are no longer needed either due to the now proper message ordering.

Updated the GenericNetowrkObjectBehaviour such that clients disable the MeshRenderer component and then have a delayed re-enabling of it upon being spawned in order to avoid the latent update in the transform's position (removes the flickering of NetworkObjects on the client).

Added the ability to make the SwitchSceneHandler have a delayed "auto switch" in order to run cycled long burn runs to assure everything is working "as expected".

Fixed an access violation in StatsDisplay ( minor thing that was bugging me).

* fix and refactor

Refactored a better client side "no blinking when spawning" solution.

Fixed an issue where everyone was sending destroy messages to everyone else.

* refactor

Added a more simplified example that uses a INetworkPrefabInstanceHandler implementation as a 1:1 NetworkObject pool.

* fix

Fixed issue with spawning:
- Sending to every client for each client within the observers list
- Making the client ownership equal to the client it was sending to

This fixes the issue with connecting more than 1 client to a host or server.