UTP: Add proper support for channels (#986)
* Correctly handle empty messages in Send

The UTPTransport.Send method would throw a NullReferenceException
exception when passed default(ArraySegment<byte>) as the message data.
Now it simply sends an empty message to its remote host (not a terribly
useful feature, but better than an exception).

* Make it possible to receive messages >1024 bytes

The fragmentation pipeline was set up to handle messages up to 6KB in
length, but the receive buffer was only 1KB long. Messages longer than
that were just ignored (although an error message was logged).

* Add proper support for channels in UTPTransport

The transport now selects an appropriate pipeline according to the
delivery level of the channel, rather than send everything on the
default unreliable pipeline.