refactor
Removing the option to send a payload when spawning an object.
The primary reason for this is that we do not save the payload data on a per NetworkObject instance basis and if a user were to use the spawn payload parameter that included data that was required to set any specific states of a NetworkObject or associated components then late joining players would not receive the payload information and would become out of sych with the game.