fix: corrected an off-by-one copying the netvar buffer into NetworkVariableDelta message
Other tests missed this because they were sending small numbers and the extra space in the receive buffer was initialized with 0s (i.e., in little endian, a 32 bit number representing 1 would be 01 00 00 00, if you lose the last byte you've lost 00, when the buffer is initialized with 0s, you happen to get the correct byte back).

Added a test to specifically exercise this.