feat
Improved notifications and detection of notifications for manual verification that they are triggering properly.
Added a SceneEventNotificationQueue test project script that is to be used with the NetworkManager to detect edge case notifications (i.e. in the middle of a scene transition).
Fixed an edge case scenario with NetworkObject.Despawn that could occur during exiting the application and the SpawnManager no longer exists.