fix, refactor, and style
re-fixing an issue with sending spawn calls to every client (n) times based on the number of clients connected (until it is re-fixed by the offender)

Started the removal of certain NetworkSceneManager event notifications that are no longer required due to the more recent SceneEvent notifications.

Removed the set active scene scene event for MS-1 (just not enough time to implement and be sure that it works perfectly)

Removed a manual test for SceneManager callbacks that won't be used any longer.

Updated XML documentation in various areas.