feat
SwitchSceneProgress changed to SceneEventProgress and is no longer a returned value when starting a scene event (load or unload).  This is still used internally to track when all clients have loaded the scene in question, but that is primarily used for SceneLoadingMode.Single.

Removed the coroutine callback from NetworkManager and placed it within the SceneEventProgress class.