refactor fix and test
Removed the client event notification as it wasn't needed.
Fixed a minor bug in scene validation that was discovered via the unit test.
Added another unit test to the NetworkSceneManagerTests that verifies the NetworkSceneManager scene validation process works and that users can control which scenes are loaded (test covers both client and server).
Updated and added XML document comments for methods and properties.