fix
Removing my tools fix as that was the wrong place to fix that issues.  The error causing issue actually came before those two tests where something was being instantiated but never destroyed.  If this fails then I might need to do the same thing with the Setup side of things as well.