fix
This fixes the scenario where you open a scene in the editor that does not have a NetworkManager instance (i.e. MainMenu in testproject) and then while still in play mode navigate to a scene that does have a NetworkManager instance, then (editor only) the ScenesInBuild asset will be created (if one doesn't exist), refreshed/repopulated, and assigned to the NetworkManager instance.