refactor and fix
Refactoring some of the test project scripts, fixing an issue in the editor where ScenesInBuild count could become 0 but not populated, slight geometry level tweaks to help prevent corner edge cases, and merging the prefab handler with the prefab pool to reduce code complexity.  (really need to merge the prefab pools together)
Also some minor tweaks to the random movement when not the owner for temporary adjustment (this needs to become velocity based).