refactor unit test check
-------------------------------------------------------------
After further consideration, the best idea for the whole unit test issue is to:
1.) Not try to detect it ( silly idea of mine )
2,) Make changes to the NetworkSceneManager's private m_ScenesLoaded property by renaming it to ScenesLoaded and making it internal.
3.) Update the one Unit test that was failing to  populate ScenesLoaded and ServerSceneHandleToClientSceneHandle tables so the SetTheSceneBeingSynchronized method could properly be tested during the NetworkObjectSceneSerializationFailure test.

It turns out my brain decided to take a "(dis)connect" day a few days too early.
No need to check for unit tests at all for this PR.
-------------------------------------------------------------
This also includes a minor adjustment for when the NetworkPrefabPool (test project script) unsubscribes from the NetworkSceneManager.OnSceneEvent to avoid a null access exception.