refactor
Took Fatih's suggestion regarding ScenesLoaded being over-complicated and now the ScenesLoaded is just a dictionary of scene handle keys and each key entry value is the Scene it is associated with.  This did reduce some of the complexity in the unloading methods and the GetAndAddNewlyLoadedSceneByName method.

Updated the NetworkObjectSceneSerializationFailure unit test to reflect the above changes.