fix and refactor
In order to validate scene loading for (n) clients and a server/host, I did add an automated way to check if a unit test is running (for now) in order to assure clients don't try to load a single mode scene.
Also merged the recent changes to the feat/MTT-820-AdditiveSceneLoading branch,  and made some adjustments to those changes in NetworkSceneManagerTests.