refactor
This is a merge refactoring due to an issue with one of the dependent branches (test/additivesceneloading) that was broken while Fatih was showing me an alternate way to handle daisy chaining branches.

Went ahead and followed the same path Fatih took with fixing the additive scene testing.

It still has an issue with the unit test, but it is close to being back to the way it was with the additional updates from PR-955.